升级核心组件

+++ app/model/level.go
@@ -29,7 +29,7 @@ type LevelAddReq struct {
 
 // 编辑职级
 type LevelEditReq struct {
-	Id     int64  `p:"id" v:"required#主键ID不能为空"`
+	Id     int  `p:"id" v:"required#主键ID不能为空"`
 	Name   string `p:"name"  v:"required#职级名称不能为空"`
 	Status int    `p:"status"    v:"required#职级状态不能为空"`
 	Sort   int    `p:"sort"  v:"required#显示顺序不能为空"`
